+++ packages/core/src/utils/deriveClassNames.test.ts
@@ -238,6 +238,7 @@ describe('deriveRuleGroupClassNames', () => {
         'shiftActions',
         'undoAction',
         'undoRedoActions',
+        'ungroup',
       ].toSorted()
     );
   });